RF-XX: remove auth router
