--- src/routes/auth/index.ts
@@ -1,9 +1,0 @@
-import express from 'express';
-
-import meRouter from './me';
-
-const router = express.Router();
-
-router.use('/me', meRouter);
-
-export default router;
