+++ src/routes/index.ts
@@ -2,16 +2,14 @@ import express from 'express';
 
 import { isUser } from 'src/middlewares/firebase';
 
-import authRouter from './auth';
 import exerciseRouter from './exercise';
-import onboardingRouter from './onboarding';
+import meRouter from './me';
 import routineRouter from './routine';
 
 const router = express.Router();
 
-router.use('/auth', isUser, authRouter);
+router.use('/me', isUser, meRouter);
 router.use('/exercise', isUser, exerciseRouter);
 router.use('/routine', isUser, routineRouter);
-router.use('/onboarding', isUser, onboardingRouter);
 
 export default router;
