+++ src/routes/me/training-preferences/utils.ts
